+++ src/ffmpegio/__init__.py
@@ -49,7 +49,7 @@ from . import ffmpegprocess
 
 from .utils.error import FFmpegError
 from .utils.concat import FFConcat
-from .filtergraph import FilterGraph
+from .filtergraph import Graph as FilterGraph
 from . import devices, ffmpegprocess, caps, probe, audio, image, video, media
 from .transcode import transcode
 from . import streams as _streams