+++ src/ffmpegio/video.py
@@ -1,5 +1,6 @@
+import warnings
 from . import ffmpegprocess, utils, configure, FFmpegError, probe, plugins, analyze
-from .utils import filter as filter_utils, log as log_utils
+from .utils import log as log_utils
 
 __all__ = ["create", "read", "write", "filter", "detect"]
 
@@ -55,87 +56,61 @@ def _run_read(
     )
 
 
-def create(
-    expr,
-    *args,
-    t_in=None,
-    pix_fmt=None,
-    vf=None,
-    progress=None,
-    show_log=None,
-    **kwargs,
-):
+def create(expr, *args, progress=None, show_log=None, **options):
     """Create a video using a source video filter
 
     :param name: name of the source filter
     :type name: str
-    :param \\*args: filter arguments
-    :type \\*args: tuple, optional
-    :param t_in: duration of the video in seconds, defaults to None
-    :type t_in: float, optional
-    :param pix_fmt_in: input pixel format if known but not specified in the ffmpeg arg dict, defaults to None
-    :type pix_fmt_in: str, optional
-    :param vf: additional video filter, defaults to None
-    :type vf: FilterGraph or str, optional
+    :param \\*args: sequential filter option arguments. Only valid for
+                    a single-filter expr, and they will overwrite the
+                    options set by expr.
+    :type \\*args: seq, optional
     :param progress: progress callback function, defaults to None
     :type progress: callable object, optional
     :param show_log: True to show FFmpeg log messages on the console,
                      defaults to None (no show/capture)
                      Ignored if stream format must be retrieved automatically.
     :type show_log: bool, optional
-    :param \\**options: filter keyword arguments
+    :param \\**options: Named filter options or FFmpeg options. Items are
+                        only considered as the filter options if expr is a
+                        single-filter graph, and take the precedents over
+                        general FFmpeg options. Append '_in' for input
+                        option names (see :doc:`options`), and '_out' for
+                        output option names if they conflict with the filter
+                        options.
     :type \\**options: dict, optional
     :return: frame rate and video data, created by `bytes_to_video` plugin hook
     :rtype: tuple[Fraction,object]
 
-    See https://ffmpeg.org/ffmpeg-filters.html#Video-Sources for available video source filters
+    ...seealso::
+      https://ffmpeg.org/ffmpeg-filters.html#Video-Sources for available
+      video source filters
 
     """
 
-    # =============  ==============================================================================
-    # filter name    description
-    # =============  ==============================================================================
-    # "color"        uniformly colored frame
-    # "allrgb"       frames of size 4096x4096 of all rgb colors
-    # "allyuv"       frames of size 4096x4096 of all yuv colors
-    # "gradients"    several gradients
-    # "mandelbrot"   Mandelbrot set fractal
-    # "mptestsrc"    various test patterns of the MPlayer test filter
-    # "life"         life pattern based on John Conway’s life game
-    # "haldclutsrc"  identity Hald CLUT
-    # "testsrc"      test video pattern, showing a color pattern
-    # "testsrc2"     another test video pattern, showing a color pattern
-    # "rgbtestsrc"   RGB test pattern useful for detecting RGB vs BGR issues
-    # "smptebars"    color bars pattern, based on the SMPTE Engineering Guideline EG 1-1990
-    # "smptehdbars"  color bars pattern, based on the SMPTE RP 219-2002
-    # "pal100bars"   a color bars pattern, based on EBU PAL recommendations with 100% color levels
-    # "pal75bars"    a color bars pattern, based on EBU PAL recommendations with 75% color levels
-    # "yuvtestsrc"   YUV test pattern. You should see a y, cb and cr stripe from top to bottom
-    # "sierpinski"   Sierpinski carpet/triangle fractal
-    # =============  ==============================================================================
+    input_options = utils.pop_extra_options(options, "_in")
+    output_options = utils.pop_extra_options(options, "_out")
 
-    url, (r_in, s_in) = filter_utils.compose_source("video", expr, *args, **kwargs)
+    url, t_, options = configure.config_input_fg(expr, args, options)
 
-    need_t = ("mandelbrot", "life")
-    if t_in is None and any((expr.startswith(f) for f in need_t)):
-        raise ValueError(f"Some sources {need_t} must have t_in specified")
+    if t_ is not None or not any(
+        a in options for a in ("t", "to", "t_in", "to_in", "frames:v", "vframes")
+    ):
+        warnings.warn(
+            "neither input nor output duration specified. this function call may hang."
+        )
+
+    options = {**options, **output_options}
 
     ffmpeg_args = configure.empty()
-    inopts = configure.add_url(ffmpeg_args, "input", url, {"f": "lavfi"})[1][1]
-    outopts = configure.add_url(ffmpeg_args, "output", "-", {})[1][1]
-
-    if t_in is not None:
-        inopts["t"] = t_in
-
-    for k, v in zip(
-        ("pix_fmt", "filter:v"),
-        (pix_fmt or "rgb24", vf),
-    ):
-        if v is not None:
-            outopts[k] = v
+    configure.add_url(ffmpeg_args, "input", url, {**input_options, "f": "lavfi"})
+    configure.add_url(ffmpeg_args, "output", "-", {**options, "f": "rawvideo"})
 
     return _run_read(
-        ffmpeg_args, progress=progress, r_in=r_in, s_in=s_in, show_log=show_log
+        ffmpeg_args,
+        pix_fmt_in=input_options.get("pix_fmt", "rgb24"),
+        progress=progress,
+        show_log=show_log,
     )
 
 
@@ -252,7 +227,6 @@ def write(
                 configure.add_url(ffmpeg_args, "input", input)
             else:
                 configure.add_url(ffmpeg_args, "input", *input)
-
 
     configure.add_url(ffmpeg_args, "output", url, options)